Add API to configure Microservices with router
It allows to set any other router (instead of RoundRobin) as default to
all ServiceCalls for particular Microservice.

Changed behavior of Routers factory - if router can't be created
because of any exception, this exception will be propagated upside
instead of returning null from factory method.

If Routers factory throws exception during instantiation of router
which is passed to @Inject annotation, this exception will be
propagated upside instead of using default configured router.